requirements.txt
gunicorn==20.1.0
Flask==2.0.3
Flask_SQLAlchemy==2.5.1
mysql_connector_repackaged==0.3.1
pandas==1.4.1
SQLAlchemy==1.4.31
mysql-connector-python==8.0.28